Completed documentation of basic functional  vc library/
